Remove other spots that did not use the internal pip version to exectue pipenv commands.
